requirements.txt
fastapi==0.73.0
uvicorn==0.17.4
motor==2.5.1
pydantic[email]
python-decouple==3.6
passlib[bcrypt]
PyJWT == 1.7.1